Added: AbstractResult class
